Move CiftiSeparateMetric from utils to interfaces.
