+++ xcp_d/workflow/restingstate.py
@@ -15,7 +15,7 @@ from niworkflows.engine.workflows import LiterateWorkflow as Workflow
 from templateflow.api import get as get_template
 
 from xcp_d.interfaces.resting_state import brainplot, computealff, surfaceReho
-from xcp_d.utils.ciftiseparatemetric import CiftiSeparateMetric
+from xcp_d.interfaces.workbench import CiftiSeparateMetric
 from xcp_d.utils.utils import fwhm2sigma
